Reservation Aggregator/requirements.txt
# Instructions: To install dependencies run: pip install -r requirements.txt
# Web Scraping
selenium

# Environment Configurations
python-dotenv

# Networking
requests
